[#49979841] Don't allow making a question mandatory for a finalised survey.
This is necessary because the other validation that enforced that the only
fields updatable are `content` and `order_number` now kicks in only if the
question is finalised.

But even if the question is non-finalised but the survey is finalised, we shouldn't
be able to make it mandatory.

+++ app/assets/javascripts/backbone/views/questions/option_view.js.coffee
@@ -62,4 +62,5 @@ class SurveyBuilder.Views.Questions.OptionView extends Backbone.View
     view.remove()
 
   limit_edit: =>
-    $(this.el).find(".delete_option").hide()
+    if @model.get("finalized")
+      $(this.el).find(".delete_option").hide()
